requirements.txt
# Core dependencies
aiogram==3.15.0
aiohttp==3.10.10

# Database (SQLite - no installation required!)
aiosqlite==0.20.0
SQLAlchemy==2.0.36
greenlet==3.1.1

# Scheduler
APScheduler==3.10.4

# Environment variables
python-dotenv==1.0.1

# Utils
python-dateutil==2.9.0
